mv kiezkassen comment column to core
