+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Comment/Comment.ts
@@ -346,6 +346,19 @@ export var adhCreateOrShowCommentListing = (
     };
 };
 
+export var commentColumnDirective = (
+    adhConfig : AdhConfig.IService
+) => {
+    return {
+        restrict: "E",
+        templateUrl: adhConfig.pkg_path + pkgLocation + "/CommentColumn.html",
+        require: "^adhMovingColumn",
+        link: (scope, element, attrs, column : AdhMovingColumns.MovingColumnController) => {
+            column.bindVariablesAndClear(scope, ["commentCloseUrl", "commentableUrl"]);
+        }
+    };
+};
+
 
 export var moduleName = "adhComment";
 
@@ -386,5 +399,6 @@ export var register = (angular) => {
                 var widget = new CommentCreate(
                     adapter, adhConfig, adhHttp, adhPermissions, adhPreliminaryNames, adhTopLevelState, $window, $q);
                 return widget.createRecursionDirective(adhRecursionHelper);
-            }]);
+            }])
+        .directive("adhCommentColumn", ["adhConfig", commentColumnDirective]);
 };
