+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Comment/CommentColumn.html
